[feature] 기본 구조 설정
