+++ packages/domain/.gitignore
