+++ packages/presentation/lib/app/wine/models/wine.freezed.dart
