--- presentation/lib/presentation.dart
@@ -1,7 +1,0 @@
-library presentation;
-
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
